Updated getDecimal(color) to not call String.format
I noticed getDecimal's call to String.format was causing a lot of slowdown, especially when running alongside Sodium/Rubidium, as those seem to make a lot of calls to biome and sky colors. Converting the RBG values to a hex string using Integer.toHexString and concatenation gets back 10 to 20 fps.